refactor(pin): Moving checkForMaintenance to AuthenticationCoordinator.
