+++ Blockchain/AppDelegate.swift
@@ -15,10 +15,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     // MARK: - Properties
 
-    // TODO: move to authentication flow
-    /// Flag used to indicate whether the device is prompting for biometric authentication.
-    @objc public private(set) var isPromptingForBiometricAuthentication = false
-
     lazy var busyView: BCFadeView? = {
         guard let windowFrame = UIApplication.shared.keyWindow?.frame else {
             return BCFadeView(frame: UIScreen.main.bounds)
@@ -100,7 +96,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationWillResignActive(_ application: UIApplication) {
         print("applicationWillResignActive")
-        if !isPromptingForBiometricAuthentication {
+        if !AuthenticationCoordinator.shared.isPromptingForBiometricAuthentication {
             showPrivacyScreen()
         }
     }
@@ -121,50 +117,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationWillTerminate(_ application: UIApplication) {
 
-    }
-
-    // MARK: - Authentication
-
-    @objc func authenticateWithBiometrics() {
-        app.pinEntryViewController.view.isUserInteractionEnabled = false
-        isPromptingForBiometricAuthentication = true
-        AuthenticationManager.shared.authenticateUsingBiometrics { authenticated, authenticationError in
-            self.isPromptingForBiometricAuthentication = false
-            if let error = authenticationError {
-                self.handleBiometricAuthenticationError(with: error)
-            }
-            DispatchQueue.main.async {
-                app.pinEntryViewController.view.isUserInteractionEnabled = true
-            }
-            if authenticated {
-                DispatchQueue.main.async {
-                    // TODO move this method to AuthenticatioCoordinator
-//                    self.showVerifyingBusyView(withTimeout: 30)
-                }
-                guard let pinKey = BlockchainSettings.App.shared.pinKey,
-                    let pin = KeychainItemWrapper.pinFromKeychain() else {
-                        self.failedToObtainValuesFromKeychain(); return
-                }
-                WalletManager.shared.wallet.apiGetPINValue(pinKey, pin: pin)
-            }
-        }
-    }
-
-    // TODO: migrate to the responsible controller that prompts for authentication
-    func handleBiometricAuthenticationError(with error: AuthenticationError) {
-        if let description = error.description {
-            let alert = UIAlertController(title: LocalizationConstants.Errors.error, message: description, preferredStyle: .alert)
-            let action = UIAlertAction(title: LocalizationConstants.ok, style: .default, handler: nil)
-            alert.addAction(action)
-            DispatchQueue.main.async {
-                UIApplication.shared.keyWindow?.rootViewController?.present(alert, animated: true, completion: nil)
-            }
-        }
-    }
-
-    // TODO: migrate to the responsible controller that prompts for authentication
-    func handlePasscodeAuthenticationError(with error: AuthenticationError) {
-        // TODO: implement handlePasscodeAuthenticationError
     }
 
     func application(_ app: UIApplication, open url: URL, options: [UIApplicationOpenURLOptionsKey: Any] = [:]) -> Bool {
@@ -225,16 +177,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func showPrivacyScreen() {
         privacyScreen?.alpha = 1
         UIApplication.shared.keyWindow?.addSubview(privacyScreen!)
-    }
-
-    func failedToObtainValuesFromKeychain() {
-        let alert = UIAlertController(title: "", message: "", preferredStyle: .alert)
-        let action = UIAlertAction(title: "", style: .cancel, handler: { _ in
-            // let app = UIApplication.shared
-            // perform suspend selector
-        })
-        alert.addAction(action)
-        UIApplication.shared.keyWindow?.rootViewController?.present(alert, animated: true, completion: nil)
     }
 
     //: These two functions are used to justify the regeneration of addresses in the swipe-to-receive screen.
