+++ Blockchain/Authentication/Models/AuthenticationCoordinator+Pin.swift
@@ -37,11 +37,7 @@ extension AuthenticationCoordinator: PEPinEntryControllerDelegate {
         //        }
         //        #endif
 
-        // TODO: migrate check for maintenance
-
-        //        dispatch_async(dispatch_get_main_queue(), ^{
-        //            [self checkForMaintenanceWithPinKey:pinKey pin:pin];
-        //        });
+        checkForMaintenance(withPinKey: pinKey, pin: pinString)
 
         self.pinViewControllerCallback = callback
     }
@@ -107,6 +103,27 @@ extension AuthenticationCoordinator: PEPinEntryControllerDelegate {
         closePinEntryView(animated: true)
     }
 
+    // MARK: - Internal
+
+    internal func showVerifyingBusyView(withTimeout seconds: Int) {
+        LoadingViewPresenter.shared.showBusyView(withLoadingText: LocalizationConstants.verifying)
+
+        // TODO: this timeout approach should be deprecated in favor of checking actual success/error responses
+        if #available(iOS 10.0, *) {
+            loginTimeout = Timer.scheduledTimer(withTimeInterval: TimeInterval(seconds), repeats: false) { [weak self] _ in
+                self?.showLoginError()
+            }
+        } else {
+            loginTimeout = Timer.scheduledTimer(
+                timeInterval: TimeInterval(seconds),
+                target: self,
+                selector: #selector(showLoginError),
+                userInfo: nil,
+                repeats: false
+            )
+        }
+    }
+
     // MARK: - Private
 
     private func reopenChangePin() {
@@ -126,25 +143,6 @@ extension AuthenticationCoordinator: PEPinEntryControllerDelegate {
         UIApplication.shared.keyWindow?.rootViewController?.view.addSubview(pinViewController.view)
 
         pinEntryViewController = pinViewController
-    }
-
-    private func showVerifyingBusyView(withTimeout seconds: Int) {
-        LoadingViewPresenter.shared.showBusyView(withLoadingText: LocalizationConstants.verifying)
-
-        // TODO: this timeout approach should be deprecated in favor of checking actual success/error responses
-        if #available(iOS 10.0, *) {
-            loginTimeout = Timer.scheduledTimer(withTimeInterval: TimeInterval(seconds), repeats: false) { [weak self] _ in
-                self?.showLoginError()
-            }
-        } else {
-            loginTimeout = Timer.scheduledTimer(
-                timeInterval: TimeInterval(seconds),
-                target: self,
-                selector: #selector(showLoginError),
-                userInfo: nil,
-                repeats: false
-            )
-        }
     }
 }
